Move getEntityStatus into a file within src/api
